kconfigs: mehlow: add 4.14
Note: BIOS does not yet enable USB 3.0, which delays the kernel boot
for a non-trivial amount of time because the USB drives freaks out
over various errors.  Things eventually get resolved, but since this
is a headless system, just disable USB for the standard config.  But,
to avoid having to play tricks to get USB support, use a USB enabled
kernel for GRUB_DEFAULT.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>