bin: git-email: Drop support for sending "indy" patches
Drop "indy" patches, the code is ugly the feature itself is a mess. The
idea with "indy" patches was to allow sending a subset of patches in a
larger series to a curated list of people, i.e. to avoid spamming people.
In practice, people generally don't care, and if they do care, i.e. the
series is enormous, then it's easier to just Cc folks on the cover letter
and send the rest only to self+lkml and point them at the lore link.

Signed-off-by: Sean Christopherson <seanjc@google.com>