bash: testing: Test more KVM module params settings when testing "everything"
Test Intel with EPT A/D, unrestricted guest, and APICv on/off (in addition
to testing with EPT on/of).

Test AMD with AVIC on/off.

Simultaneously update the modprobe wrappers to default to avic being
enabled, and add 'mpkr' to probe without any "defaults".  Trying to repro a
failure gets quite confusing when features are off by default.

Signed-off-by: Sean Christopherson <seanjc@google.com>